HBX-1647: Move class 'org.hibernate.tool.hbm2x.POJOExporter' to 'org.hibernate.tool.internal.export.pojo.POJOExporter'
